requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.31.0
virtualenv==20.27.1
Werkzeug==3.1.3
setuptools==75.8.0
supertools==1.0.1
wheel==0.45.1
